pages/admin-dashboard: rearchitecting giving resources their own page
